refactor based on code review feedback
